Move TileCache classes to separate file
